requirements.txt
streamlit
biopython
matplotlib
requests
numpy
stmol
py3Dmol
ipython_genutils
ipywidgets
fpdf